Moved code from main.cpp and app.h to app.cpp and app_init.h
